% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_labels}
\alias{cell_labels}
\title{Cell labels}
\format{
A vector of length equal to the number of cells and in the same order as it is on the single cell object.
}
\usage{
cell_labels
}
\description{
A character vector or factor representing the cell type labels for each cell.
}
\examples{
data(cell_labels)
head(cell_labels)
}
\keyword{datasets}
