% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells_groundtruth}
\alias{cells_groundtruth}
\title{Cell groundtruth}
\format{
Matrix with samples as rows and cell types as columns
}
\usage{
cells_groundtruth
}
\description{
A matrix with the cell type proportions of the single cell object.
}
\examples{
data(cells_groundtruth)
head(cells_groundtruth)
}
\keyword{datasets}
