% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{computeDWLS}
\alias{computeDWLS}
\title{Computes DWLS}
\usage{
computeDWLS(TPM_matrix, signature_file, name_signature)
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (genes symbols as rows and samples as columns).}

\item{signature_file}{The signature file to use.}

\item{name_signature}{Signature name to set for the deconvolution results.}
}
\value{
A matrix with cell abundance deconvolve with DWLS
}
\description{
Computes DWLS
}
\keyword{internal}
