% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{computeMOMF_parallel}
\alias{computeMOMF_parallel}
\title{Compute MOMF in parallel across multiple signatures}
\usage{
computeMOMF_parallel(TPM_matrix, sc_object, signatures, workers)
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (genes symbols as rows and samples as columns).}

\item{sc_object}{A matrix with the counts from scRNAseq object (genes as rows and cells as columns)}

\item{signatures}{Path where signatures files are located}

\item{workers}{Number of processes available to run on parallel.}
}
\value{
A matrix with cell abundance deconvolve with MOMF
}
\description{
Compute MOMF in parallel across multiple signatures
}
\keyword{internal}
