% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{correlation}
\alias{correlation}
\title{Perform pairwise correlation across all features}
\usage{
correlation(data, corr_type = "spearman")
}
\arguments{
\item{data}{Matrix with features to correlate}

\item{corr_type}{Correlation type whether "spearman" or "pearson".}
}
\value{
Dataframe containing all significant correlations (pvalue < 0.05)
}
\description{
Perform pairwise correlation across all features
}
