% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deconv_bulk}
\alias{deconv_bulk}
\title{Bulk deconvolution matrix}
\format{
Matrix with samples as rows and deconvolution features as columns
}
\usage{
deconv_bulk
}
\description{
A matrix with the cell type deconvolution features obtained from compute.deconvolution() from the bulk RNAseq of Mariathasan et al. (2018)
}
\examples{
data(deconv_bulk)
head(deconv_bulk)
}
\keyword{datasets}
