% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deconvolution}
\alias{deconvolution}
\title{Deconvolution matrix}
\format{
Matrix with samples as rows and deconvolution features as columns
}
\usage{
deconvolution
}
\description{
A matrix with the cell type deconvolution features obtained from compute.deconvolution() from the pseudobulk matrix
}
\examples{
data(deconvolution)
head(deconvolution)
}
\keyword{datasets}
