% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{find.maximum.iteration}
\alias{find.maximum.iteration}
\title{Find maximum iteration from subgroups}
\usage{
find.maximum.iteration(cells.groups)
}
\arguments{
\item{cells.groups}{Cell groups corresponding to a specific cell type.}
}
\value{
Maximum subgroupping iteration
}
\description{
Find maximum iteration from subgroups
}
