% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacells_data}
\alias{metacells_data}
\title{Metacells data}
\format{
Matrix with samples as columns and genes names (SYMBOL) as rows
}
\source{
Senosain et al. (2023), doi: 10.1158/2767-9764.CRC-22-0373. PMID: 37501683; PMCID: PMC10370362
}
\usage{
metacells_data
}
\description{
A gene expression matrix corresponding to the single cell object, in this case it corresponds to the reduced single cell object (metacells), not the original.
}
\examples{
data(metacells_data)
head(metacells_data)
}
\keyword{datasets}
