% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacells_metadata}
\alias{metacells_metadata}
\title{Metacells metadata}
\format{
Matrix with samples as rows and meta information as columns
}
\source{
Senosain et al. (2023), doi: 10.1158/2767-9764.CRC-22-0373. PMID: 37501683; PMCID: PMC10370362
}
\usage{
metacells_metadata
}
\description{
A dataframe corresponding to the metadata of the single cell object, in this case it corresponds to the reduced single cell object (metacells). It needs to include the columns specified as cell_labels and samples_labels.
}
\examples{
data(metacells_metadata)
head(metacells_metadata)
}
\keyword{datasets}
