% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multideconv-package.R
\docType{package}
\name{multideconv-package}
\alias{multideconv}
\alias{multideconv-package}
\title{multideconv: Integrative Pipeline for Cell Type Deconvolution from Bulk RNAseq using First and Second Generation Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package that reduces dimensionality and eliminates redundancy through unsupervised filtering and iterative correlation analyses. Built on top of existing frameworks like immunedeconv (Sturm et al. 2019) and omnideconv (Dietrich et al. 2024), multideconv harmonizes outputs across methods to identify robust cell subgroups and mitigate signature-driven heterogeneity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://verapancaldilab.github.io/multideconv/}
  \item \url{https://github.com/VeraPancaldiLab/multideconv}
}

}
\author{
\strong{Maintainer}: Marcelo Hurtado \email{marcelo.hurtado@inserm.fr} (\href{https://orcid.org/0009-0004-6712-0864}{ORCID})

Authors:
\itemize{
  \item Vera Pancaldi \email{vera.pancaldi@inserm.fr} (\href{https://orcid.org/0000-0002-7433-624X}{ORCID})
}

}
\keyword{internal}
