% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulk}
\format{
Matrix with genes as rows and samples as columns
}
\usage{
pseudobulk
}
\description{
A gene expression matrix after aggregation of the single cell object (pseudobulk).
}
\examples{
data(pseudobulk)
head(pseudobulk)
}
\keyword{datasets}
