% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_counts}
\alias{raw_counts}
\title{Raw counts}
\format{
Matrix with genes as rows and samples as columns
}
\source{
Mariathasan et al. (2018), doi: https://doi.org/10.1038/nature25501
}
\usage{
raw_counts
}
\description{
Raw gene expression matrix from bulk RNAseq.
}
\examples{
data(raw_counts)
head(raw_counts)
}
\keyword{datasets}
