% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_labels}
\alias{sample_labels}
\title{Sample labels}
\format{
A vector of length equal to the number of cells and in the same order as it is on the single cell object.
}
\usage{
sample_labels
}
\description{
A character vector or factor representing the sample labels for each cell.
}
\examples{
data(sample_labels)
head(sample_labels)
}
\keyword{datasets}
