% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{standardize_celltype_colnames}
\alias{standardize_celltype_colnames}
\title{Standardize Cell Type Column Names}
\usage{
standardize_celltype_colnames(mat)
}
\arguments{
\item{mat}{A matrix with cell type data.}
}
\value{
A matrix with standardized cell type column names.
}
\description{
This function standardizes the column names of a matrix containing cell type data.
}
\examples{
mat <- matrix(rnorm(30), nrow = 10, ncol = 3)
colnames(mat) <- c("Macrophage_M0", "Macrophage_M1", "Macrophage_M2")
standardized_mat <- multideconv:::standardize_celltype_colnames(mat)


}
