% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{compute.cell.types}
\alias{compute.cell.types}
\title{Cell types split from deconvolution}
\usage{
compute.cell.types(data, cells_extra = NULL)
}
\arguments{
\item{data}{A matrix with the deconvolution results}

\item{cells_extra}{A string specifying the cells names to consider and that are not including in the nomenclature of multideconv (see Readme)}
}
\value{
A list containing:
\itemize{
\item A sublist with each cell type features as an element recover from the different signatures
\item Discarded cell types (this will happen if the cell types are not supported. See the READme for more information about this)
}
}
\description{
Cell types split from deconvolution
}
\examples{

data("deconvolution")

cells_types = compute.cell.types(deconvolution)
cells = cells_types[[1]]
cells_discarded = cells_types[[2]]
cells_types = compute.cell.types(deconvolution, cells_extra = c("mesenchymal", "basophils"))

}
