% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{compute.deconvolution.preprocessing}
\alias{compute.deconvolution.preprocessing}
\title{Compute deconvolution preprocessing}
\usage{
compute.deconvolution.preprocessing(deconv)
}
\arguments{
\item{deconv}{A dataframe with the unprocessed deconvolution features}
}
\value{
A matrix of the preprocessed deconvolution features with fixed and consistent names across the different methods and signatures following the nomenclature specified in multideconv (see Readme)
}
\description{
Give consistent names and patterns following the method_signature_cell structure to the deconvolution features
}
\examples{

data("deconvolution")

deconvolution = compute.deconvolution.preprocessing(deconvolution)

}
