% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{compute_methods_variable_signature}
\alias{compute_methods_variable_signature}
\title{Compute deconvolution methods with variable signatures}
\usage{
compute_methods_variable_signature(
  TPM_matrix,
  signatures,
  algos = c("CBSX", "Epidish", "DeconRNASeq", "DWLS"),
  exclude = NULL,
  cbsx.name,
  cbsx.token,
  doParallel = FALSE,
  workers = NULL,
  sc_obj = NULL
)
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (samples as columns and genes symbols as rows)}

\item{signatures}{A path with a directory where signatures are located}

\item{algos}{A character vector with the methods to compute (Default methods are CBSX, Epidish, DeconRNASeq and DWLS)}

\item{exclude}{(Optional) A character vector with the signature to exclude}

\item{cbsx.name}{Cibersortx credential mail if CBSX will be run}

\item{cbsx.token}{Cibersortx credential token if CBSX will be run}

\item{doParallel}{Boolean value to specify if DWLS and CBSX should run in parallel (default is False)}

\item{workers}{Number of worker process to run during parallelization (default is NULL)}

\item{sc_obj}{A matrix with the counts from scRNAseq object (genes as rows and cells as columns) to run MOMF method. If NULL, MOMF is ignored.}
}
\value{
A matrix with the deconvolution features corresponding to all combinations of methods-signatures specified
}
\description{
Compute deconvolution methods with variable signatures
}
\examples{

data("raw_counts")

TPM_matrix = ADImpute::NormalizeTPM(raw_counts)

deconvolution = compute_methods_variable_signature(TPM_matrix, algos = c("CBSX", "Epidish"),
                                                   cbsx.name = "XXX", cbsx.token = "XXX")

}
