% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{compute_subgroups}
\alias{compute_subgroups}
\title{Compute deconvolution subgroups}
\usage{
compute_subgroups(deconvolution, thres_corr, file_name)
}
\arguments{
\item{deconvolution}{A matrix with unprocessed cell deconvolution results}

\item{thres_corr}{A numeric value with the minimum correlation allowed to group cell deconvolution features}

\item{file_name}{Base name for subgroup}
}
\value{
A list containing
\itemize{
\item A matrix with the processed deconvolution features
\item Cell subgroups obtained by linear correlation
\item Cell subgroups obtained by proportionality correlation
\item Discard cell features either because of low variance or high zero number
}
}
\description{
Compute deconvolution subgroups
}
\examples{

data("deconvolution")
B_cells = deconvolution[,grep("B.cells", colnames(deconvolution))]

x = compute_subgroups(B_cells, file_name = "B.cells", thres_corr = 0.7)

}
