% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{computeCBSX}
\alias{computeCBSX}
\title{Computes CibersortX (CBSX) using one signature}
\usage{
computeCBSX(TPM_matrix, signature_file, name, password, name_signature)
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (genes symbols as rows and samples as columns).}

\item{signature_file}{The signature file to use.}

\item{name}{Credential email for running CibersortX.}

\item{password}{Credential token for running CibersortX.}

\item{name_signature}{Signature name to set for the deconvolution results.}
}
\value{
A matrix with cell abundance deconvolve with CBSX
}
\description{
Computes CibersortX (CBSX) using one signature
}
\examples{

data("raw_counts")

TPM_matrix = ADImpute::NormalizeTPM(raw_counts)

cbsx <- computeCBSX(TPM_matrix, signature, cbsx.name, cbsx.token, signature_name)

}
