% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{computeMOMF}
\alias{computeMOMF}
\title{Computes MOMF}
\usage{
computeMOMF(TPM_matrix, sc_object, signature_file, name_signature)
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (genes symbols as rows and samples as columns).}

\item{sc_object}{A matrix with the counts from scRNAseq object (genes as rows and cells as columns)}

\item{signature_file}{The signature file to use.}

\item{name_signature}{Signature name to set for the deconvolution results.}
}
\value{
A matrix with cell abundance deconvolve with MOMF
}
\description{
Computes MOMF
}
