% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{computeQuantiseq}
\alias{computeQuantiseq}
\title{Computes QuanTIseq}
\usage{
computeQuantiseq(TPM_matrix, name_signature = "TIL10")
}
\arguments{
\item{TPM_matrix}{A matrix with TPM normalized counts (genes symbols as rows and samples as columns).}

\item{name_signature}{Name used to tag output columns with the signature source.}
}
\value{
A matrix with cell abundance deconvolve with QuanTIseq
}
\description{
Computes QuanTIseq
}
