% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{corr_subgroups}
\alias{corr_subgroups}
\title{Perform pairwise correlation across all features}
\usage{
corr_subgroups(data, corr_type = "spearman", batch = NULL)
}
\arguments{
\item{data}{Matrix with features to correlate}

\item{corr_type}{Correlation type whether "spearman" or "pearson".}

\item{batch}{Optional batch covariate used to compute partial correlations.}
}
\value{
Dataframe containing all significant correlations (pvalue < 0.05)
}
\description{
Perform pairwise correlation across all features
}
