% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{create_metacells}
\alias{create_metacells}
\title{Create meta-cells from a single cell object}
\usage{
create_metacells(
  sc_object,
  labels_column,
  samples_column,
  exclude_cells = NULL,
  min_cells = 50,
  k = 15,
  max_shared = 15,
  n_workers = 4,
  min_meta = 10
)
}
\arguments{
\item{sc_object}{A matrix with the counts from scRNAseq object (genes as rows and cells as columns)}

\item{labels_column}{A character vector with the cell labels (need to be of the same order as in the sc_object)}

\item{samples_column}{A character vector with the samples labels (need to be of the same order as in the sc_object)}

\item{exclude_cells}{Cell types to discard from metacell algorithm.}

\item{min_cells}{The minimum number of cells in a particular grouping to construct metacells.}

\item{k}{Number of nearest neighbors to aggregate for KNN algorithm.}

\item{max_shared}{The maximum number of cells to be shared across two metacells.}

\item{n_workers}{Number of cores to use for paralellization.}

\item{min_meta}{Minimum number of metacells allowed. Below this number, metacells of this cell type will be discarded.}
}
\value{
A list with two elements:
\itemize{
\item The metacell count matrix (genes as rownames and cells as columns)
\item The metadata matrix corresponding to the metacell object
}
}
\description{
Create meta-cells from a single cell object
}
\examples{

data("metacells_data")
data("sample_labels")
data("cell_labels")


metacells = create_metacells(metacells, cell_labels, sample_labels, exclude_cells = NULL,
                             min_cells = 100, k = 30, max_shared = 15, n_workers = 4)

}
