% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{create_sc_pseudobulk}
\alias{create_sc_pseudobulk}
\title{Create pseudo bulk from single cell object}
\usage{
create_sc_pseudobulk(
  sc_obj,
  cells_labels,
  sample_labels,
  normalized = TRUE,
  file_name
)
}
\arguments{
\item{sc_obj}{A Seurat single cell object}

\item{cells_labels}{A character vector with the cell labels (need to be of the same order as in the sc_obj)}

\item{sample_labels}{A character vector with the samples labels (need to be of the same order as in the sc_obj)}

\item{normalized}{Whether pseudobulk should be or not TPM normalized}

\item{file_name}{A string specifying the name of the .csv pseudobulk saved in Results/}
}
\value{
A gene count matrix (genes as rows and samples as columns)
}
\description{
Create pseudo bulk from single cell object
}
