% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{create_sc_signatures}
\alias{create_sc_signatures}
\title{Create cell type signatures from scRNAseq}
\usage{
create_sc_signatures(
  sc_obj,
  sc_metadata,
  cells_labels,
  sample_labels,
  credentials.mail = NULL,
  credentials.token = NULL,
  bulk_rna = NULL,
  cell_markers = NULL,
  name_signature = NULL
)
}
\arguments{
\item{sc_obj}{A matrix with the counts from scRNAseq object (genes as rows and cells as columns)}

\item{sc_metadata}{Dataframe with metadata from the single cell object. The matrix should include the columns cell_label and sample_label.}

\item{cells_labels}{A character vector with the cell labels (need to be of the same order as in the sc_object)}

\item{sample_labels}{A character vector with the samples labels (need to be of the same order as in the sc_object)}

\item{credentials.mail}{(Optional) Credential email for running CibersortX. If not provided, cibersortX method will not be run.}

\item{credentials.token}{(Optional) Credential token for running CibersortX. If not provided, cibersortX method will not be run.}

\item{bulk_rna}{A matrix of bulk data. Rows are genes, columns are samples. This is needed for MOMF method, if not given the method will not be run.}

\item{cell_markers}{Named list with the genes markers names as Symbol per cell types to be used to create the signature using the BSeq-SC method. If NULL, the method will be ignored during the signature creation.}

\item{name_signature}{A string indicating the signature name. This will be added as a suffix in each method (e.g. CBSX_name_signature, DWLS_name_signature)}
}
\value{
A list containing the cell signatures per method. Signatures are directly saved in signatures/ folders, these will be used to run deconvolution.
}
\description{
Create cell type signatures from scRNAseq
}
\examples{

data("metacells_data")
data("metacells_metadata")
data("sample_labels")
data("cell_labels")

signatures = create_sc_signatures(metacells_data, metacells_metadata, cell_labels, sample_labels,
                                  credentials.mail = "XXX", credentials.token = "XXX",
                                  bulk_rna, name_signature = "Test")

}
