% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{remove_low_variance}
\alias{remove_low_variance}
\title{Remove low variance deconvolution features}
\usage{
remove_low_variance(data, var_quantile = 0.25)
}
\arguments{
\item{data}{Deconvolution features}

\item{var_quantile}{Quantile threshold below which features are discarded.}
}
\value{
A list containing
\itemize{
\item Deconvolution matrix after removal of low variance.
\item Discarded low variance features.
}
}
\description{
Remove low variance deconvolution features
}
