% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{removeCorrelatedFeatures}
\alias{removeCorrelatedFeatures}
\title{Remove high correlated cell deconvolution features}
\usage{
removeCorrelatedFeatures(
  data,
  threshold,
  name,
  n_seed,
  corr_method = "spearman",
  batch = NULL
)
}
\arguments{
\item{data}{Deconvolution matrix}

\item{threshold}{Threshold for defined high correlated features}

\item{name}{Cell type name corresponding to the given matrix in 'data'}

\item{n_seed}{Seed to ensure reproducibility regarding the choice of the feature.}

\item{corr_method}{Correlation type whether "spearman" or "pearson".}

\item{batch}{Optional batch covariate used to compute partial correlations.}
}
\value{
A list containing
\itemize{
\item Deconvolution matrix with only one deconvolution feature per high-correlated pair.
\item Highly correlated features found
\item Cell type name
}
}
\description{
If two deconvolution features within a specific cell type are found to be highly correlated, one feature is kept randomly for further analysis.
}
