% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_deconvolution.R
\name{replicate_deconvolution_subgroups}
\alias{replicate_deconvolution_subgroups}
\title{Replicate deconvolution subgroups in a new dataset}
\usage{
replicate_deconvolution_subgroups(deconv_res, deconvolution_test)
}
\arguments{
\item{deconv_res}{A list containing results from the deconvolution process, including:
\itemize{
\item{\code{Deconvolution subgroups composition}: a list of subgroup feature names per iteration}
\item{\code{Deconvolution matrix}: the original deconvolution result used to determine relevant features}
}}

\item{deconvolution_test}{A data.frame or matrix of deconvolution results (e.g., from another cohort)}
}
\value{
A data.frame with the projected subgroup features proportions
}
\description{
Reconstructs and applies deconvolution subgroup signatures based on a previous decomposition.
}
