% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subgroups}
\alias{subgroups}
\title{Cell subgroups}
\format{
A list with the cell subgroups
}
\usage{
subgroups
}
\description{
Cell subgroups composition
}
\examples{
data(subgroups)
subgroups[[1]]
}
\keyword{datasets}
